Extract cast_to_list to utils module
